Generates the json config for the plugin with one monitor point for each antenna
Instead of one monitor point for each values returned by the utility
module, it returns one monitor point for each antenna.

The reason is that the tulity module actually returns a set of bits
and it makes sens to send all of them at once.
In this way the plugins also injects monitor points that we could
need in the future.